+++ homeassistant/components/tasmota/const.py
@@ -1,7 +1,4 @@
 """Constants used by multiple Tasmota modules."""
-ATTR_DISCOVERY_PAYLOAD = "discovery_payload"
-ATTR_DISCOVERY_TOPIC = "discovery_topic"
-
 CONF_DISCOVERY_PREFIX = "discovery_prefix"
 
 DEFAULT_PREFIX = "tasmota/discovery"
