+++ homeassistant/components/tasmota/discovery.py
@@ -4,19 +4,13 @@ import logging
 
 from hatasmota.const import CONF_RELAY
 from hatasmota.discovery import (
+    TasmotaDiscovery,
     get_device_config as tasmota_get_device_config,
     get_entities_for_platform as tasmota_get_entities_for_platform,
     get_entity as tasmota_get_entity,
     has_entities_with_platform as tasmota_has_entities_with_platform,
-    subscribe_discovery_topic,
 )
 
-from homeassistant.components import mqtt
-from homeassistant.components.mqtt.subscription import (
-    async_subscribe_topics,
-    async_unsubscribe_topics,
-)
-from homeassistant.core import callback
 from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.helpers.typing import HomeAssistantType
 
@@ -31,7 +25,6 @@ SUPPORTED_COMPONENTS = {
 ALREADY_DISCOVERED = "tasmota_discovered_components"
 CONFIG_ENTRY_IS_SETUP = "tasmota_config_entry_is_setup"
 DATA_CONFIG_ENTRY_LOCK = "tasmota_config_entry_lock"
-DISCOVERY_UNSUBSCRIBE = "tasmota_discovery_unsubscribe"
 TASMOTA_DISCOVERY_DEVICE = "tasmota_discovery_device"
 TASMOTA_DISCOVERY_ENTITY_NEW = "tasmota_discovery_entity_new_{}"
 TASMOTA_DISCOVERY_ENTITY_UPDATED = "tasmota_discovery_entity_updated_{}_{}_{}"
@@ -48,7 +41,7 @@ def set_discovery_hash(hass, discovery_hash):
 
 
 async def async_start(
-    hass: HomeAssistantType, discovery_topic, config_entry=None
+    hass: HomeAssistantType, discovery_topic, config_entry, tasmota_mqtt
 ) -> bool:
     """Start MQTT Discovery."""
 
@@ -110,25 +103,10 @@ async def async_start(
                     _LOGGER.info("Adding new entity: %s %s", component, discovery_hash)
                     hass.data[ALREADY_DISCOVERED][discovery_hash] = None
 
-                    def _publish(*args):
-                        mqtt.async_publish(hass, *args)
-
-                    async def _subscribe_topics(sub_state, topics):
-                        # Mark message handlers as callback
-                        for topic in topics.values():
-                            if "msg_callback" in topic:
-                                topic["msg_callback"] = callback(topic["msg_callback"])
-                        return await async_subscribe_topics(hass, sub_state, topics)
-
-                    async def _unsubscribe_topics(sub_state):
-                        return await async_unsubscribe_topics(hass, sub_state)
-
                     tasmota_entity = tasmota_get_entity(
-                        tasmota_entity_config, component_key
+                        tasmota_entity_config, component_key, tasmota_mqtt
                     )
-                    tasmota_entity.set_mqtt_callbacks(
-                        _publish, _subscribe_topics, _unsubscribe_topics
-                    )
+
                     async_dispatcher_send(
                         hass,
                         TASMOTA_DISCOVERY_ENTITY_NEW.format(component),
@@ -139,13 +117,7 @@ async def async_start(
     hass.data[DATA_CONFIG_ENTRY_LOCK] = asyncio.Lock()
     hass.data[CONFIG_ENTRY_IS_SETUP] = set()
 
-    async def _subscribe(topic, msg_callback, *args):
-        hass.data[DISCOVERY_UNSUBSCRIBE] = await mqtt.async_subscribe(
-            hass, topic, msg_callback, *args
-        )
-
-    await subscribe_discovery_topic(
-        discovery_topic, async_device_discovered, _subscribe
-    )
+    tasmota_discovery = TasmotaDiscovery(discovery_topic, tasmota_mqtt)
+    await tasmota_discovery.start_discovery(async_device_discovered)
 
     return True
