+++ homeassistant/components/tasmota/manifest.json
@@ -3,7 +3,7 @@
   "name": "Tasmota",
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/tasmota",
-  "requirements": ["hatasmota==0.0.5"],
+  "requirements": ["hatasmota==0.0.6"],
   "dependencies": ["mqtt"],
   "codeowners": ["@emontnemery"]
 }
