+++ requirements_all.txt
@@ -729,7 +729,7 @@ hass-nabucasa==0.37.0
 hass_splunk==0.1.1
 
 # homeassistant.components.tasmota
-hatasmota==0.0.5
+hatasmota==0.0.6
 
 # homeassistant.components.jewish_calendar
 hdate==0.9.5
