+++ requirements_test_all.txt
@@ -361,7 +361,7 @@ hangups==0.4.11
 hass-nabucasa==0.37.0
 
 # homeassistant.components.tasmota
-hatasmota==0.0.5
+hatasmota==0.0.6
 
 # homeassistant.components.jewish_calendar
 hdate==0.9.5
